# Backend/api/laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Final Grado API",
        "description": "Documentacion de la API para el Proyecto Final de Grado",
        "contact": {
            "name": "Alejandro Solar, David Kovacs",
            "email": "alejandrosolarruiz@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/generos": {
            "get": {
                "tags": [
                    "Genero"
                ],
                "summary": "Listar todos los géneros paginados",
                "operationId": "3f481aeb2388ba232907c09ac43e7f7d",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "registrosPorPagina",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de géneros paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Genero"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Genero"
                ],
                "summary": "Crear un nuevo género",
                "operationId": "e79f2f2d12fe7a22ed1d0754903b3c9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "description": "Imagen en base64 (opcional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Género creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genero"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/generos/{id}": {
            "get": {
                "tags": [
                    "Genero"
                ],
                "summary": "Obtener un género por ID",
                "operationId": "1fa17e1f1e1166a23829328ca0997986",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Género encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genero"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Género no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Genero"
                ],
                "summary": "Actualizar un género existente",
                "operationId": "b5523d9efe666d307c1511addc56367a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "description": "Imagen en base64 (opcional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Género actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genero"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Género no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genero"
                ],
                "summary": "Eliminar un género",
                "operationId": "910530f740dd29219c9520e0b0fb2464",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Género eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Género eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Género no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Genero": {
                "properties": {
                    "id_genero": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Acción"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}